+++ src/Plutarch/Internal/Builtin.hs
@@ -10,6 +10,9 @@ module Plutarch.Internal.Builtin (
   PBool (..),
   PString (..),
   PByteString (..),
+  PBLS12_381_G1_Element (..),
+  PBLS12_381_G2_Element (..),
+  PBLS12_381_MlResult (..),
 
   -- * Functions
 
@@ -55,6 +58,25 @@ module Plutarch.Internal.Builtin (
   pbuiltinEqualsByteString,
   pbuiltinLessThanByteString,
   pbuiltinLessThanEqualsByteString,
+
+  -- ** BLS
+  pbuiltinBls12_381_G1_add,
+  pbuiltinBls12_381_G1_scalarMul,
+  pbuiltinBls12_381_G1_neg,
+  pbuiltinBls12_381_G1_compress,
+  pbuiltinBls12_381_G1_uncompress,
+  pbuiltinBls12_381_G1_hashToGroup,
+  pbuiltinBls12_381_G1_equal,
+  pbuiltinBls12_381_G2_add,
+  pbuiltinBls12_381_G2_scalarMul,
+  pbuiltinBls12_381_G2_neg,
+  pbuiltinBls12_381_G2_compress,
+  pbuiltinBls12_381_G2_uncompress,
+  pbuiltinBls12_381_G2_hashToGroup,
+  pbuiltinBls12_381_G2_equal,
+  pbuiltinBls12_381_millerLoop,
+  pbuiltinBls12_381_mulMlResult,
+  pbuiltinBls12_381_finalVerify,
 
   -- ** Other
   pfix,
@@ -111,6 +133,9 @@ import Plutarch.Lift (
   pconstant,
  )
 import PlutusCore qualified as PLC
+import PlutusCore.Crypto.BLS12_381.G1 qualified as BLS12_381_G1
+import PlutusCore.Crypto.BLS12_381.G2 qualified as BLS12_381_G2
+import PlutusCore.Crypto.BLS12_381.Pairing qualified as Pairing
 
 -- An arbitrary term whose type is unknown.
 --
@@ -262,6 +287,93 @@ instance Semigroup (Term s PByteString) where
 instance Monoid (Term s PByteString) where
   {-# INLINEABLE mempty #-}
   mempty = pconstant ""
+
+{- | A point on the BLS12-381 G1 curve.
+
+@since WIP
+-}
+newtype PBLS12_381_G1_Element (s :: S)
+  = PBLS12_381_G1_Element (Term s POpaque)
+  deriving stock
+    ( -- | @since WIP
+      Generic
+    )
+  deriving anyclass
+    ( -- | @since WIP
+      PlutusType
+    )
+
+-- | @since WIP
+instance DerivePlutusType PBLS12_381_G1_Element where
+  type DPTStrat _ = PlutusTypeNewtype
+
+-- | @since WIP
+instance PUnsafeLiftDecl PBLS12_381_G1_Element where
+  type PLifted PBLS12_381_G1_Element = BLS12_381_G1.Element
+
+-- | @since WIP
+deriving via
+  (DerivePConstantDirect BLS12_381_G1.Element PBLS12_381_G1_Element)
+  instance
+    PConstantDecl BLS12_381_G1.Element
+
+{- | A point on the BLS12-381 G2 curve.
+
+@since WIP
+-}
+newtype PBLS12_381_G2_Element (s :: S)
+  = PBLS12_381_G2_Element (Term s POpaque)
+  deriving stock
+    ( -- | @since WIP
+      Generic
+    )
+  deriving anyclass
+    ( -- | @since WIP
+      PlutusType
+    )
+
+-- | @since WIP
+instance DerivePlutusType PBLS12_381_G2_Element where
+  type DPTStrat _ = PlutusTypeNewtype
+
+-- | @since WIP
+instance PUnsafeLiftDecl PBLS12_381_G2_Element where
+  type PLifted PBLS12_381_G2_Element = BLS12_381_G2.Element
+
+-- | @since WIP
+deriving via
+  (DerivePConstantDirect BLS12_381_G2.Element PBLS12_381_G2_Element)
+  instance
+    PConstantDecl BLS12_381_G2.Element
+
+{- | The result of a Miller loop in BLS12-381 pairing.
+
+@since WIP
+-}
+newtype PBLS12_381_MlResult (s :: S)
+  = PBLS12_381_MlResult (Term s POpaque)
+  deriving stock
+    ( -- | @since WIP
+      Generic
+    )
+  deriving anyclass
+    ( -- | @since WIP
+      PlutusType
+    )
+
+-- | @since WIP
+instance DerivePlutusType PBLS12_381_MlResult where
+  type DPTStrat _ = PlutusTypeNewtype
+
+-- | @since WIP
+instance PUnsafeLiftDecl PBLS12_381_MlResult where
+  type PLifted PBLS12_381_MlResult = Pairing.MlResult
+
+-- | @since WIP
+deriving via
+  (DerivePConstantDirect Pairing.MlResult PBLS12_381_MlResult)
+  instance
+    PConstantDecl Pairing.MlResult
 
 {- | Forget the type of a term.
 
@@ -594,3 +706,105 @@ punsafeDowncast ::
   Term s (PInner a) ->
   Term s a
 punsafeDowncast = punsafeCoerce
+
+-- | @since WIP
+pbuiltinBls12_381_G1_add ::
+  forall (s :: S).
+  Term s (PBLS12_381_G1_Element :--> PBLS12_381_G1_Element :--> PBLS12_381_G1_Element)
+pbuiltinBls12_381_G1_add = punsafeBuiltin PLC.Bls12_381_G1_add
+
+-- | @since WIP
+pbuiltinBls12_381_G1_scalarMul ::
+  forall (s :: S).
+  Term s (PInteger :--> PBLS12_381_G1_Element :--> PBLS12_381_G1_Element)
+pbuiltinBls12_381_G1_scalarMul = punsafeBuiltin PLC.Bls12_381_G1_scalarMul
+
+-- | @since WIP
+pbuiltinBls12_381_G1_neg ::
+  forall (s :: S).
+  Term s (PBLS12_381_G1_Element :--> PBLS12_381_G1_Element)
+pbuiltinBls12_381_G1_neg = punsafeBuiltin PLC.Bls12_381_G1_neg
+
+-- | @since WIP
+pbuiltinBls12_381_G1_compress ::
+  forall (s :: S).
+  Term s (PBLS12_381_G1_Element :--> PByteString)
+pbuiltinBls12_381_G1_compress = punsafeBuiltin PLC.Bls12_381_G1_compress
+
+-- | @since WIP
+pbuiltinBls12_381_G1_uncompress ::
+  forall (s :: S).
+  Term s (PByteString :--> PBLS12_381_G1_Element)
+pbuiltinBls12_381_G1_uncompress = punsafeBuiltin PLC.Bls12_381_G1_uncompress
+
+-- | @since WIP
+pbuiltinBls12_381_G1_hashToGroup ::
+  forall (s :: S).
+  Term s (PByteString :--> PByteString :--> PBLS12_381_G1_Element)
+pbuiltinBls12_381_G1_hashToGroup = punsafeBuiltin PLC.Bls12_381_G1_hashToGroup
+
+-- | @since WIP
+pbuiltinBls12_381_G1_equal ::
+  forall (s :: S).
+  Term s (PBLS12_381_G1_Element :--> PBLS12_381_G1_Element :--> PBool)
+pbuiltinBls12_381_G1_equal = punsafeBuiltin PLC.Bls12_381_G1_equal
+
+-- | @since WIP
+pbuiltinBls12_381_G2_add ::
+  forall (s :: S).
+  Term s (PBLS12_381_G2_Element :--> PBLS12_381_G2_Element :--> PBLS12_381_G2_Element)
+pbuiltinBls12_381_G2_add = punsafeBuiltin PLC.Bls12_381_G2_add
+
+-- | @since WIP
+pbuiltinBls12_381_G2_scalarMul ::
+  forall (s :: S).
+  Term s (PInteger :--> PBLS12_381_G2_Element :--> PBLS12_381_G2_Element)
+pbuiltinBls12_381_G2_scalarMul = punsafeBuiltin PLC.Bls12_381_G2_scalarMul
+
+-- | @since WIP
+pbuiltinBls12_381_G2_neg ::
+  forall (s :: S).
+  Term s (PBLS12_381_G2_Element :--> PBLS12_381_G2_Element)
+pbuiltinBls12_381_G2_neg = punsafeBuiltin PLC.Bls12_381_G2_neg
+
+-- | @since WIP
+pbuiltinBls12_381_G2_compress ::
+  forall (s :: S).
+  Term s (PBLS12_381_G2_Element :--> PByteString)
+pbuiltinBls12_381_G2_compress = punsafeBuiltin PLC.Bls12_381_G2_compress
+
+-- | @since WIP
+pbuiltinBls12_381_G2_uncompress ::
+  forall (s :: S).
+  Term s (PByteString :--> PBLS12_381_G2_Element)
+pbuiltinBls12_381_G2_uncompress = punsafeBuiltin PLC.Bls12_381_G2_uncompress
+
+-- | @since WIP
+pbuiltinBls12_381_G2_hashToGroup ::
+  forall (s :: S).
+  Term s (PByteString :--> PByteString :--> PBLS12_381_G2_Element)
+pbuiltinBls12_381_G2_hashToGroup = punsafeBuiltin PLC.Bls12_381_G2_hashToGroup
+
+-- | @since WIP
+pbuiltinBls12_381_G2_equal ::
+  forall (s :: S).
+  Term s (PBLS12_381_G2_Element :--> PBLS12_381_G2_Element :--> PBool)
+pbuiltinBls12_381_G2_equal = punsafeBuiltin PLC.Bls12_381_G2_equal
+
+-- | @since WIP
+pbuiltinBls12_381_millerLoop ::
+  forall (s :: S).
+  Term s (PBLS12_381_G1_Element :--> PBLS12_381_G2_Element :--> PBLS12_381_MlResult)
+pbuiltinBls12_381_millerLoop = punsafeBuiltin PLC.Bls12_381_millerLoop
+
+-- | @since WIP
+pbuiltinBls12_381_mulMlResult ::
+  forall (s :: S).
+  Term s (PBLS12_381_MlResult :--> PBLS12_381_MlResult :--> PBLS12_381_MlResult)
+pbuiltinBls12_381_mulMlResult = punsafeBuiltin PLC.Bls12_381_mulMlResult
+
+-- | @since WIP
+pbuiltinBls12_381_finalVerify ::
+  forall (s :: S).
+  Term s (PBLS12_381_MlResult :--> PBLS12_381_MlResult :--> PBool)
+pbuiltinBls12_381_finalVerify = punsafeBuiltin PLC.Bls12_381_finalVerify