+++ CHANGELOG.md
@@ -58,6 +58,7 @@
 * `Plutarch.Num`, `Plutarch.Integer`, as they were unified into a new module
 * `Plutarch.Unsafe`, as it no longer contained anything useful
 * `Plutarch` module, as it was generally useless
+* `Plutarch.BLS`, as it's now part of `Plutarch.Internal.Builtin`
 
 ### Fixed